refactor: extract agent studio git contracts
