+++ apps/desktop/src/components/features/agents/agent-studio-git-panel/pull-rebase-dialog.tsx
@@ -1,6 +1,6 @@
 import { ArrowDown } from "lucide-react";
 import { memo, type ReactElement } from "react";
-import type { AgentStudioPendingPullRebase } from "@/pages/agents/use-agent-studio-git-actions";
+import type { AgentStudioPendingPullRebase } from "@/features/agent-studio-git";
 import { INLINE_CODE_CLASS_NAME } from "./constants";
 import { GitConfirmationDialog } from "./git-confirmation-dialog";
