+++ apps/desktop/src/components/features/agents/agent-studio-git-panel/rebase-conflict-dialog.tsx
@@ -8,7 +8,7 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
-import type { AgentStudioRebaseConflict } from "@/pages/agents/use-agent-studio-git-actions";
+import type { AgentStudioRebaseConflict } from "@/features/agent-studio-git";
 import { INLINE_CODE_CLASS_NAME } from "./constants";
 import { RebaseConflictActions, type RebaseConflictActionsModel } from "./rebase-conflict-actions";
