+++ apps/desktop/src/features/agent-studio-git/index.ts
@@ -1,0 +1,9 @@
+export type {
+  AgentStudioPendingForcePush,
+  AgentStudioPendingPullRebase,
+  AgentStudioRebaseConflict,
+  AgentStudioRebaseConflictAction,
+  AgentStudioRebaseConflictOperation,
+  DiffDataState,
+  DiffScope,
+} from "./contracts";
