+++ apps/desktop/src/pages/agents/use-agent-studio-diff-data.test.tsx
@@ -77,6 +77,12 @@ const gitGetWorktreeStatusSummaryMock = mock(
   },
 );
 
+mock.module("@/lib/host-client", () => ({
+  hostClient: {
+    runsList: runsListMock,
+  },
+}));
+
 mock.module("@/state/operations/host", () => ({
   host: {
     runsList: runsListMock,
